update db impact

+++ src/main/java/com/tujuhsembilan/wrcore/dto/TaskDTO.java
@@ -1,5 +1,7 @@
 package com.tujuhsembilan.wrcore.dto;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Id;
 
 import lombok.*;
@@ -10,14 +12,14 @@ import lombok.*;
 @Builder
 public class TaskDTO {
 
-    @Id
-    private Long taskId;
-    private Long backlogId;
-    private Long workingReportId;
-    private Long categoryCodeId;
-    private String taskItem;
-    private int duration;
-    private boolean isOvertime;
-    private String filePath;
+  @Id
+  private Long taskId;
+  private Long backlogId;
+  private Long workingReportId;
+  private Long categoryCodeId;
+  private String taskItem;
+  private BigDecimal duration;
+  private boolean isOvertime;
+  private String filePath;
 
 }
